refactor: extract MockAioRpcError to shared conftest module
Deduplicate MockAioRpcError class that was identically defined in both
grpc_test.py and core_test.py. The shared class now lives in conftest.py
at the jumpstarter package root and is imported by both test files.

Generated-By: Forge/20260601_132018_678661_8c9cb822
